assignment4/requirements.txt
git+https://github.com/Revo1999/vrashelper
polars
pillow
tqdm
torch
facenet-pytorch